attept to build alveo-u280-hbm soc
